refactor(ui): move UI components within components folder
