+++ packages/gboost-ui/package.json
@@ -32,6 +32,7 @@
     "@mantine/hooks": "^3.6.14",
     "@radix-ui/colors": "^0.1.8",
     "@radix-ui/react-dialog": "^0.1.7",
+    "@radix-ui/react-portal": "^0.1.4",
     "@stitches/react": "^1.2.8",
     "gboost-common": "workspace:^0.5.1",
     "graphql": "^15.8.0",
