+++ packages/gboost-ui/src/QueryTable/Pagination.tsx
@@ -1,9 +1,8 @@
 import { Button, SelectField, Text } from "@aws-amplify/ui-react";
 import { ReactElement } from "react";
 import { MdChevronLeft, MdChevronRight } from "react-icons/md";
-import { Box } from "../Box.js";
+import { Box, styled } from "../index.js";
 import type { CSS } from "@stitches/react";
-import { styled } from "../stitches.config.js";
 
 const StyledButton = styled(Button, {
   "&:hover": { backgroundColor: "$primary3" },
