+++ packages/gboost-ui/src/QueryTable/TableHeaderCell.tsx
@@ -1,10 +1,9 @@
 import { ReactElement, useCallback, useRef } from "react";
 import { Button, TableCell } from "@aws-amplify/ui-react";
-import { config, styled } from "../stitches.config.js";
 import * as Stitches from "@stitches/react";
 import { MdArrowDownward, MdArrowUpward, MdFilterList } from "react-icons/md";
 import { Column, Sort } from "./QueryTable.js";
-import { Box } from "../Box.js";
+import { Box, config, styled } from "../index.js";
 import { RefObject } from "react";
 
 const StyledTableCell = styled(TableCell, {
