+++ packages/gboost-ui/src/components/Dialog.tsx
@@ -1,9 +1,8 @@
 import { ReactElement } from "react";
 import { Heading, Icon } from "@aws-amplify/ui-react";
-import { keyframes, styled } from "./stitches.config.js";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
-import { Box } from "./Box.js";
 import { MdClose } from "react-icons/md";
+import { keyframes, styled, Box } from "../index.js";
 
 const overlayShow = keyframes({
   "0%": { opacity: 0 },
